third_party/WebKit/LayoutTests/svg/css/opacity-not-supporting-percentage.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
    description("Test opacity properties combined with percentage.");

    shouldBeFalse('CSS.supports("stroke-opacity", "0%")');
    shouldBeFalse('CSS.supports("stroke-opacity", "100%")');

    shouldBeFalse('CSS.supports("flood-opacity", "0%")');
    shouldBeFalse('CSS.supports("flood-opacity", "100%")');

    shouldBeFalse('CSS.supports("stop-opacity", "0%")');
    shouldBeFalse('CSS.supports("stop-opacity", "100%")');

    shouldBeFalse('CSS.supports("opacity", "0%")');
    shouldBeFalse('CSS.supports("opacity", "100%")');
</script>
</body>
</html>
