third_party/WebKit/LayoutTests/svg/css/opacity-not-supporting-percentage-expected.txt

Test opacity properties combined with percentage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS CSS.supports("stroke-opacity", "0%") is false
PASS CSS.supports("stroke-opacity", "100%") is false
PASS CSS.supports("flood-opacity", "0%") is false
PASS CSS.supports("flood-opacity", "100%") is false
PASS CSS.supports("stop-opacity", "0%") is false
PASS CSS.supports("stop-opacity", "100%") is false
PASS CSS.supports("opacity", "0%") is false
PASS CSS.supports("opacity", "100%") is false
PASS successfullyParsed is true

TEST COMPLETE
